moved the interfaces and instruments around
