+++ Starfish/grid_tools/grid_tools.py
@@ -107,849 +107,6 @@ def determine_chunk_log(wl, wl_min, wl_max):
     return ind
 
 
-class RawGridInterface:
-    '''
-    A base class to handle interfacing with synthetic spectral libraries.
-
-    :param name: name of the spectral library
-    :type name: string
-    :param param_names: the names of the parameters (dimensions) of the grid
-    :type param_names: list
-    :param points: the grid points at which
-        spectra exist (assumes grid is square, not ragged, meaning that every combination
-        of parameters specified exists in the grid).
-    :type points: list of numpy arrays
-    :param air: Are the wavelengths measured in air?
-    :type air: bool
-    :param wl_range: the starting and ending wavelength ranges of the grid to
-        truncate to.
-    :type wl_range: list of len 2 [min, max]
-    :param base: path to the root of the files on disk.
-    :type base: string
-
-    '''
-
-    def __init__(self, name, param_names, points, air=True, wl_range=[3000, 13000], base=None):
-        self.name = name
-
-        self.param_names = param_names
-        self.points = points
-
-        self.air = air
-        self.wl_range = wl_range
-        self.base = base
-
-    def check_params(self, parameters):
-        '''
-        Determine if the specified parameters are allowed in the grid.
-
-        :param parameters: parameter set to check
-        :type parameters: np.array
-
-        :raises C.GridError: if the parameter values are outside of the grid bounds
-
-        '''
-        if len(parameters) != len(self.param_names):
-            raise ValueError("Length of given parameters ({}) does not match length of grid parameters ({})".format(
-                len(parameters), len(self.param_names)))
-
-        for param, ppoints in zip(parameters, self.points):
-            if param not in ppoints:
-                raise ValueError("{} not in the grid points {}".format(param, ppoints))
-
-    def load_flux(self, parameters, norm=True):
-        '''
-        Load the synthetic flux from the disk and  :meth:`check_params`
-
-        :param parameters: stellar parameters describing a spectrum
-        :type parameters: np.array
-        :param norm: normalize the spectrum to solar luminosity?
-        :type norm: bool
-
-         .. note::
-
-            This method is designed to be extended by the inheriting class
-        '''
-        raise NotImplementedError("`load_flux` is abstract and must be implemented by subclasses")
-
-
-class PHOENIXGridInterface(RawGridInterface):
-    '''
-    An Interface to the PHOENIX/Husser synthetic library.
-
-    '''
-
-    def __init__(self, air=True, wl_range=[3000, 54000], base=config.grid["raw_path"]):
-
-        super().__init__(name="PHOENIX",
-                         param_names=["temp", "logg", "Z", "alpha"],
-                         points=[
-                             np.array([2300, 2400, 2500, 2600, 2700, 2800, 2900, 3000, 3100, 3200,
-                                       3300, 3400, 3500, 3600, 3700, 3800, 3900, 4000, 4100, 4200, 4300, 4400,
-                                       4500, 4600, 4700, 4800, 4900, 5000, 5100, 5200, 5300, 5400, 5500, 5600,
-                                       5700, 5800, 5900, 6000, 6100, 6200, 6300, 6400, 6500, 6600, 6700, 6800,
-                                       6900, 7000, 7200, 7400, 7600, 7800, 8000, 8200, 8400, 8600, 8800, 9000,
-                                       9200, 9400, 9600, 9800, 10000, 10200, 10400, 10600, 10800, 11000, 11200,
-                                       11400, 11600, 11800, 12000]),
-                             np.arange(0.0, 6.1, 0.5),
-                             np.arange(-2., 1.1, 0.5),
-                             np.array([-0.2, 0.0, 0.2, 0.4, 0.6, 0.8])],
-                         air=air, wl_range=wl_range, base=base)  # wl_range used to be [2999, 13001]
-
-        self.par_dicts = [None,
-                          None,
-                          {-2 : "-2.0", -1.5: "-1.5", -1: '-1.0', -0.5: '-0.5',
-                           0.0: '-0.0', 0.5: '+0.5', 1: '+1.0'},
-                          {-0.4: ".Alpha=-0.40", -0.2: ".Alpha=-0.20",
-                           0.0 : "", 0.2: ".Alpha=+0.20", 0.4: ".Alpha=+0.40",
-                           0.6 : ".Alpha=+0.60", 0.8: ".Alpha=+0.80"}]
-
-        # if air is true, convert the normally vacuum file to air wls.
-        try:
-            wl_filename = os.path.join(self.base, "WAVE_PHOENIX-ACES-AGSS-COND-2011.fits")
-            w_full = fits.getdata(wl_filename)
-        except:
-            raise ValueError("Wavelength file improperly specified.")
-
-        if self.air:
-            self.wl_full = vacuum_to_air(w_full)
-        else:
-            self.wl_full = w_full
-
-        self.ind = (self.wl_full >= self.wl_range[0]) & (self.wl_full <= self.wl_range[1])
-        self.wl = self.wl_full[self.ind]
-        self.rname = os.path.join(self.base, "Z{2:}{3:}/lte{0:0>5.0f}-{1:.2f}{2:}{3:}" \
-                                             ".PHOENIX-ACES-AGSS-COND-2011-HiRes.fits")
-
-    def load_flux(self, parameters, norm=True):
-        '''
-       Load just the flux and header information.
-
-       :param parameters: stellar parameters
-       :type parameters: np.array
-       :param norm: normalize the spectrum to solar luminosity?
-       :type norm: bool
-
-       :raises C.GridError: if the file cannot be found on disk.
-
-       :returns: tuple (flux_array, header_dict)
-
-       '''
-        self.check_params(parameters)  # Check to make sure that the keys are
-        # allowed and that the values are in the grid
-
-        # Create a list of the parameters to be fed to the format string
-        # optionally replacing arguments using the dictionaries, if the formatting
-        # of a certain parameter is tricky
-        str_parameters = []
-        for param, par_dict in zip(parameters, self.par_dicts):
-            if par_dict is None:
-                str_parameters.append(param)
-            else:
-                str_parameters.append(par_dict[param])
-
-        fname = self.rname.format(*str_parameters)
-
-        # Still need to check that file is in the grid, otherwise raise a C.GridError
-        # Read all metadata in from the FITS header, and append to spectrum
-        try:
-            flux_file = fits.open(fname)
-            f = flux_file[0].data
-            hdr = flux_file[0].header
-            flux_file.close()
-        except:
-            raise ValueError("{} is not on disk.".format(fname))
-
-        # If we want to normalize the spectra, we must do it now since later we won't have the full EM range
-        if norm:
-            f *= 1e-8  # convert from erg/cm^2/s/cm to erg/cm^2/s/A
-            F_bol = trapz(f, self.wl_full)
-            f = f * (C.F_sun / F_bol)  # bolometric luminosity is always 1 L_sun
-
-        # Add temp, logg, Z, alpha, norm to the metadata
-        header = {}
-        header["norm"] = norm
-        header["air"] = self.air
-        # Keep only the relevant PHOENIX keywords, which start with PHX
-        for key, value in hdr.items():
-            if key[:3] == "PHX":
-                header[key] = value
-
-        return (f[self.ind], header)
-
-
-class PHOENIXGridInterfaceNoAlpha(PHOENIXGridInterface):
-    '''
-    An Interface to the PHOENIX/Husser synthetic library.
-
-    '''
-
-    def __init__(self, air=True, wl_range=[3000, 54000],
-                 base=config.grid["raw_path"]):
-        # Initialize according to the regular PHOENIX values
-        super().__init__(air=air, wl_range=wl_range, base=base)
-
-        # Now override parameters to exclude alpha
-        self.param_names = ["temp", "logg", "Z"]
-        self.points = [
-            np.array([2300, 2400, 2500, 2600, 2700, 2800, 2900, 3000, 3100, 3200,
-                      3300, 3400, 3500, 3600, 3700, 3800, 3900, 4000, 4100, 4200, 4300, 4400,
-                      4500, 4600, 4700, 4800, 4900, 5000, 5100, 5200, 5300, 5400, 5500, 5600,
-                      5700, 5800, 5900, 6000, 6100, 6200, 6300, 6400, 6500, 6600, 6700, 6800,
-                      6900, 7000, 7200, 7400, 7600, 7800, 8000, 8200, 8400, 8600, 8800, 9000,
-                      9200, 9400, 9600, 9800, 10000, 10200, 10400, 10600, 10800, 11000, 11200,
-                      11400, 11600, 11800, 12000]),
-            np.arange(0.0, 6.1, 0.5),
-            np.arange(-2., 1.1, 0.5)]
-
-        self.par_dicts = [None,
-                          None,
-                          {-2 : "-2.0", -1.5: "-1.5", -1: '-1.0', -0.5: '-0.5',
-                           0.0: '-0.0', 0.5: '+0.5', 1: '+1.0'}]
-
-        self.rname = os.path.join(self.base, "Z{2:}/lte{0:0>5.0f}-{1:.2f}{2:}" \
-                                             ".PHOENIX-ACES-AGSS-COND-2011-HiRes.fits")
-
-
-class PHOENIXGridInterfaceNoAlphaNoFE(PHOENIXGridInterface):
-    '''
-    An Interface to the PHOENIX/Husser synthetic library that disregards [Fe/H] as well as alpha.
-
-    '''
-
-    def __init__(self, air=True, wl_range=[3000, 54000],
-                 base=config.grid['raw_path']):
-        # Initialize according to the regular PHOENIX values
-        super().__init__(air=air, wl_range=wl_range, base=base)
-
-        # Now override parameters to exclude alpha
-        self.param_names = ["temp", "logg"]
-        self.points = [
-            np.array([2300, 2400, 2500, 2600, 2700, 2800, 2900, 3000, 3100, 3200,
-                      3300, 3400, 3500, 3600, 3700, 3800, 3900, 4000, 4100, 4200, 4300, 4400,
-                      4500, 4600, 4700, 4800, 4900, 5000, 5100, 5200, 5300, 5400, 5500, 5600,
-                      5700, 5800, 5900, 6000, 6100, 6200, 6300, 6400, 6500, 6600, 6700, 6800,
-                      6900, 7000, 7200, 7400, 7600, 7800, 8000, 8200, 8400, 8600, 8800, 9000,
-                      9200, 9400, 9600, 9800, 10000, 10200, 10400, 10600, 10800, 11000, 11200,
-                      11400, 11600, 11800, 12000]),
-            np.arange(0.0, 6.1, 0.5)]
-
-        self.par_dicts = [None, None]
-
-        self.rname = os.path.join(self.base + "Z-0.0/lte{0:0>5.0f}-{1:.2f}-0.0" \
-                                              ".PHOENIX-ACES-AGSS-COND-2011-HiRes.fits")
-
-
-class KuruczGridInterface(RawGridInterface):
-    '''Kurucz grid interface.
-
-    Spectra are stored in ``f_nu`` in a filename like
-    ``t03500g00m25ap00k2v070z1i00.fits``, ``ap00`` means zero alpha enhancement,
-    and ``k2`` is the microturbulence, while ``z1`` is the macroturbulence.
-    These particular values are roughly the ones appropriate for the Sun.
-    '''
-
-    def __init__(self, air=True, wl_range=[5000, 5400], base=config.grid["raw_path"]):
-        super().__init__(name="Kurucz",
-                         param_names=["temp", "logg", "Z"],
-                         points=[np.arange(3500, 9751, 250),
-                                 np.arange(0.0, 5.1, 0.5),
-                                 np.array([-2.5, -2.0, -1.5, -1.0, -0.5, 0.0, 0.5])],
-                         air=air, wl_range=wl_range, base=base)
-
-        self.par_dicts = [None, None,
-                          {-2.5: "m25", -2.0: "m20", -1.5: "m15", -1.0: "m10", -0.5: "m05", 0.0: "p00", 0.5: "p05"}]
-
-        # Convert to f_lam and average to 1, or leave in f_nu?
-        self.rname = os.path.join(base, "t{0:0>5.0f}/g{1:0>2.0f}/t{0:0>5.0f}g{1:0>2.0f}{2}ap00k2v000z1i00.fits")
-        self.wl_full = np.load(os.path.join(base, "kurucz_raw_wl.npy"))
-        self.ind = (self.wl_full >= self.wl_range[0]) & (self.wl_full <= self.wl_range[1])
-        self.wl = self.wl_full[self.ind]
-
-    def load_flux(self, parameters, norm=True):
-        '''
-        Load a the flux and header information.
-
-        :param parameters: stellar parameters
-        :type parameters: dict
-        :param norm: normalize the spectrum to solar luminosity?
-        :type norm: bool
-        
-        :raises C.GridError: if the file cannot be found on disk.
-
-        :returns: tuple (flux_array, header_dict)
-
-        '''
-        self.check_params(parameters)
-
-        str_parameters = []
-        for param, par_dict in zip(parameters, self.par_dicts):
-            if par_dict is None:
-                str_parameters.append(param)
-            else:
-                str_parameters.append(par_dict[param])
-
-        # Multiply logg by 10
-        str_parameters[1] *= 10
-
-        fname = self.rname.format(*str_parameters)
-
-        # Still need to check that file is in the grid, otherwise raise a C.GridError
-        # Read all metadata in from the FITS header, and append to spectrum
-        try:
-            flux_file = fits.open(fname)
-            f = flux_file[0].data
-            hdr = flux_file[0].header
-            flux_file.close()
-        except:
-            raise ValueError("{} is not on disk.".format(fname))
-
-        # We cannot normalize the spectra, since we don't have a full wl range, so instead we set the average
-        # flux to be 1
-
-        # Also, we should convert from f_nu to f_lam
-        if norm:
-            f *= C.c_ang / self.wl ** 2  # Convert from f_nu to f_lambda
-            f /= np.average(f)  # divide by the mean flux, so avg(f) = 1
-
-        # Add temp, logg, Z, norm to the metadata
-        header = {}
-        header["norm"] = norm
-        header["air"] = self.air
-        # Keep the relevant keywords
-        for key, value in hdr.items():
-            header[key] = value
-
-        return (f[self.ind], header)
-
-
-class BTSettlGridInterface(RawGridInterface):
-    '''BTSettl grid interface. Unlike the PHOENIX and Kurucz grids, the
-    individual files of the BTSettl grid do not always have the same wavelength
-    sampling. Therefore, each call of :meth:`load_flux` will interpolate the
-    flux onto a LogLambda spaced grid that ranges between `wl_range` and has a
-    velocity spacing of 0.08 km/s or better.
-
-    If you have a choice, it's probably easier to use the Husser PHOENIX grid.
-    '''
-
-    def __init__(self, air=True, wl_range=[2999, 13000], base="libraries/raw/BTSettl/"):
-        super().__init__(name="BTSettl",
-                         points={"temp" : np.arange(3000, 7001, 100),
-                                 "logg" : np.arange(2.5, 5.6, 0.5),
-                                 "Z"    : np.arange(-0.5, 0.6, 0.5),
-                                 "alpha": np.array([0.0])},
-                         air=air, wl_range=wl_range, base=base)
-
-        # Normalize to 1 solar luminosity?
-        self.rname = os.path.join(self.base, "CIFIST2011/M{Z:}/lte{temp:0>3.0f}-{logg:.1f}{Z:}.BT-Settl.spec.7.bz2")
-        # self.Z_dict = {-2:"-2.0", -1.5:"-1.5", -1:'-1.0', -0.5:'-0.5', 0.0: '-0.0', 0.5: '+0.5', 1: '+1.0'}
-        self.Z_dict = {-0.5: '-0.5a+0.2', 0.0: '-0.0a+0.0', 0.5: '+0.5a0.0'}
-
-        wl_dict = create_log_lam_grid(wl_start=self.wl_range[0], wl_end=self.wl_range[1], min_vc=0.08 / C.c_kms)
-        self.wl = wl_dict['wl']
-
-    def load_flux(self, parameters, norm=True):
-        '''
-        Because of the crazy format of the BTSettl, we need to sort the wl to make sure
-        everything is unique, and we're not screwing ourselves with the spline.
-        
-        :param parameters: stellar parameters
-        :type parameters: dict
-        :param norm: normalize the spectrum to solar luminosity?
-        :type norm: bool
-
-        '''
-
-        super().load_flux(
-            parameters)  # Check to make sure that the keys are allowed and that the values are in the grid
-
-        str_parameters = parameters.copy()
-
-        # Rewrite Z
-        Z = parameters["Z"]
-        str_parameters["Z"] = self.Z_dict[Z]
-
-        # Multiply temp by 0.01
-        str_parameters["temp"] = 0.01 * parameters['temp']
-
-        fname = self.rname.format(**str_parameters)
-        file = bz2.BZ2File(fname, 'r')
-
-        lines = file.readlines()
-        strlines = [line.decode('utf-8') for line in lines]
-        file.close()
-
-        data = ascii.read(strlines, col_starts=[0, 13], col_ends=[12, 25], Reader=ascii.FixedWidthNoHeader)
-        wl = data['col1']
-        fl_str = data['col2']
-
-        fl = idl_float(fl_str)  # convert because of "D" exponent, unreadable in Python
-        fl = 10 ** (fl - 8.)  # now in ergs/cm^2/s/A
-
-        # "Clean" the wl and flux points. Remove duplicates, sort in increasing wl
-        wl, ind = np.unique(wl, return_index=True)
-        fl = fl[ind]
-
-        if norm:
-            F_bol = trapz(fl, wl)
-            fl = fl * (C.F_sun / F_bol)
-            # the bolometric luminosity is always 1 L_sun
-
-        # truncate the spectrum to the wl range of interest
-        # at this step, make the range a little more so that the next stage of
-        # spline interpolation is properly in bounds
-        ind = (wl >= (self.wl_range[0] - 50.)) & (wl <= (self.wl_range[1] + 50.))
-        wl = wl[ind]
-        fl = fl[ind]
-
-        if self.air:
-            # Shift the wl that correspond to the raw spectrum
-            wl = vacuum_to_air(wl)
-
-        # Now interpolate wl, fl onto self.wl
-        interp = InterpolatedUnivariateSpline(wl, fl, k=5)
-        fl_interp = interp(self.wl)
-
-        return fl_interp
-
-
-class CIFISTGridInterface(RawGridInterface):
-    '''CIFIST grid interface, grid available here: https://phoenix.ens-lyon.fr/Grids/BT-Settl/CIFIST2011_2015/FITS/.
-    Unlike the PHOENIX and Kurucz grids, the
-    individual files of the BTSettl grid do not always have the same wavelength
-    sampling. Therefore, each call of :meth:`load_flux` will interpolate the
-    flux onto a LogLambda spaced grid that ranges between `wl_range` and has a
-    velocity spacing of 0.08 km/s or better.
-
-    If you have a choice, it's probably easier to use the Husser PHOENIX grid.
-    '''
-
-    def __init__(self, air=True, wl_range=[3000, 13000], base=config.grid["raw_path"]):
-        super().__init__(name="CIFIST",
-                         points=[np.concatenate((np.arange(1200, 2351, 50), np.arange(2400, 7001, 100)), axis=0),
-                                 np.arange(2.5, 5.6, 0.5)],
-                         param_names=["temp", "logg"],
-                         air=air, wl_range=wl_range, base=base)
-
-        self.par_dicts = [None, None]
-        self.rname = os.path.join(self.base, "lte{0:0>5.1f}-{1:.1f}-0.0a+0.0.BT-Settl.spec.fits.gz")
-
-        wl_dict = create_log_lam_grid(dv=0.08, wl_start=self.wl_range[0], wl_end=self.wl_range[1])
-        self.wl = wl_dict['wl']
-
-    def load_flux(self, parameters, norm=True):
-        '''
-        Because of the crazy format of the BTSettl, we need to sort the wl to make sure
-        everything is unique, and we're not screwing ourselves with the spline.
-        
-        :param parameters: stellar parameters
-        :type parameters: dict
-        :param norm: normalize the spectrum to solar luminosity?
-        :type norm: bool
-        
-        '''
-
-        self.check_params(parameters)
-
-        str_parameters = []
-        for param, par_dict in zip(parameters, self.par_dicts):
-            if par_dict is None:
-                str_parameters.append(param)
-            else:
-                str_parameters.append(par_dict[param])
-
-        # Multiply temp by 0.01
-        str_parameters[0] = 0.01 * parameters[0]
-
-        fname = self.rname.format(*str_parameters)
-
-        # Still need to check that file is in the grid, otherwise raise a C.GridError
-        # Read all metadata in from the FITS header, and append to spectrum
-        try:
-            flux_file = fits.open(fname)
-            data = flux_file[1].data
-            hdr = flux_file[1].header
-
-            wl = data["Wavelength"] * 1e4  # [Convert to angstroms]
-            fl = data["Flux"]
-
-            flux_file.close()
-        except OSError:
-            raise C.GridError("{} is not on disk.".format(fname))
-
-        # "Clean" the wl and flux points. Remove duplicates, sort in increasing wl
-        wl, ind = np.unique(wl, return_index=True)
-        fl = fl[ind]
-
-        if norm:
-            F_bol = trapz(fl, wl)
-            fl = fl * (C.F_sun / F_bol)
-            # the bolometric luminosity is always 1 L_sun
-
-        # truncate the spectrum to the wl range of interest
-        # at this step, make the range a little more so that the next stage of
-        # spline interpolation is properly in bounds
-        ind = (wl >= (self.wl_range[0] - 50.)) & (wl <= (self.wl_range[1] + 50.))
-        wl = wl[ind]
-        fl = fl[ind]
-
-        if self.air:
-            # Shift the wl that correspond to the raw spectrum
-            wl = vacuum_to_air(wl)
-
-        # Now interpolate wl, fl onto self.wl
-        interp = InterpolatedUnivariateSpline(wl, fl, k=5)
-        fl_interp = interp(self.wl)
-
-        # Add temp, logg, Z, norm to the metadata
-        header = {}
-        header["norm"] = norm
-        header["air"] = self.air
-        # Keep the relevant keywords
-        for key, value in hdr.items():
-            header[key] = value
-
-        return (fl_interp, header)
-
-
-class HDF5Creator:
-    '''
-    Create a HDF5 grid to store all of the spectra from a RawGridInterface,
-    along with metadata.
-
-    '''
-
-    def __init__(self, GridInterface, filename, Instrument, ranges=None,
-                 key_name=config.grid["key_name"], vsinis=None):
-        '''
-        :param GridInterface: :obj:`RawGridInterface` object or subclass thereof
-            to access raw spectra on disk.
-        :param filename: where to create the HDF5 file. Suffix ``*.hdf5`` recommended.
-        :param Instrument: the instrument to convolve/truncate the grid. If you
-            want a high-res grid, use the NullInstrument.
-        :param ranges: lower and upper limits for each stellar parameter,
-            in order to truncate the number of spectra in the grid.
-        :type ranges: dict of keywords mapped to 2-tuples
-        :param key_name: formatting string that has keys for each of the parameter
-            names to translate into a hash-able string.
-        :type key_name: string
-
-        This object is designed to be run in serial.
-        '''
-
-        if ranges is None:
-            # Programatically define each range to be (-np.inf, np.inf)
-            ranges = []
-            for par in config.grid["parname"]:
-                ranges.append([-np.inf, np.inf])
-
-        self.GridInterface = GridInterface
-        self.filename = os.path.expandvars(filename)  # only store the name to the HDF5 file, because
-        # otherwise the object cannot be parallelized
-        self.Instrument = Instrument
-
-        # The flux formatting key will always have alpha in the name, regardless
-        # of whether or not the library uses it as a parameter.
-        self.key_name = key_name
-
-        # Take only those points of the GridInterface that fall within the ranges specified
-        self.points = []
-
-        # We know which subset we want, so use these.
-        for i, (low, high) in enumerate(ranges):
-            valid_points = self.GridInterface.points[i]
-            ind = (valid_points >= low) & (valid_points <= high)
-            self.points.append(valid_points[ind])
-            # Note that at this point, this is just the grid points that fall within the rectangular
-            # bounds set by ranges. If the raw library is actually irregular (e.g. CIFIST),
-            # then self.points will contain points that don't actually exist in the raw library.
-
-        # the raw wl from the spectral library
-        self.wl_native = self.GridInterface.wl  # raw grid
-        self.dv_native = calculate_dv(self.wl_native)
-
-        self.hdf5 = h5py.File(self.filename, "w")
-        self.hdf5.attrs["grid_name"] = GridInterface.name
-        self.hdf5.flux_group = self.hdf5.create_group("flux")
-        self.hdf5.flux_group.attrs["unit"] = "erg/cm^2/s/A"
-
-        # We'll need a few wavelength grids
-        # 1. The original synthetic grid: ``self.wl_native``
-        # 2. A finely spaced log-lambda grid respecting the ``dv`` of
-        #   ``self.wl_native``, onto which we can interpolate the flux values
-        #   in preperation of the FFT: ``self.wl_FFT``
-        # [ DO FFT ]
-        # 3. A log-lambda spaced grid onto which we can downsample the result
-        #   of the FFT, spaced with a ``dv`` such that we respect the remaining
-        #   Fourier modes: ``self.wl_final``
-
-        # There are three ranges to consider when wanting to make a grid:
-        # 1. The full range of the synthetic library
-        # 2. The full range of the instrument/dataset
-        # 3. The range specified by the user in config.yaml
-        # For speed reasons, we will always truncate to to wl_range. If either
-        # the synthetic library or the instrument library is smaller than this range,
-        # raise an error.
-
-        # inst_min, inst_max = self.Instrument.wl_range
-        wl_min, wl_max = config.grid["wl_range"]
-        buffer = config.grid["buffer"]  # [AA]
-        wl_min -= buffer
-        wl_max += buffer
-
-        # If the raw synthetic grid doesn't span the full range of the user
-        # specified grid, raise an error.
-        # Instead, let's choose the maximum limit of the synthetic grid?
-        if (self.wl_native[0] > wl_min) or (self.wl_native[-1] < wl_max):
-            print(
-                "Synthetic grid does not encapsulate chosen wl_range in config.yaml, truncating new grid to extent of synthetic grid, {}, {}".format(
-                    self.wl_native[0], self.wl_native[-1]))
-            wl_min, wl_max = self.wl_native[0], self.wl_native[-1]
-
-        # Calculate wl_FFT
-        # use the dv that preserves the native quality of the raw PHOENIX grid
-        wl_dict = create_log_lam_grid(self.dv_native, wl_min, wl_max)
-        self.wl_FFT = wl_dict["wl"]
-        self.dv_FFT = calculate_dv_dict(wl_dict)
-
-        print("FFT grid stretches from {} to {}".format(self.wl_FFT[0], self.wl_FFT[-1]))
-        print("wl_FFT dv is {} km/s".format(self.dv_FFT))
-
-        # The Fourier coordinate
-        self.ss = rfftfreq(len(self.wl_FFT), d=self.dv_FFT)
-
-        # The instrumental taper
-        sigma = self.Instrument.FWHM / 2.35  # in km/s
-        # Instrumentally broaden the spectrum by multiplying with a Gaussian in Fourier space
-        self.taper = np.exp(-2 * (np.pi ** 2) * (sigma ** 2) * (self.ss ** 2))
-
-        self.ss[0] = 0.01  # junk so we don't get a divide by zero error
-
-        # The final wavelength grid, onto which we will interpolate the
-        # Fourier filtered wavelengths, is part of the Instrument object
-        dv_temp = self.Instrument.FWHM / self.Instrument.oversampling
-        wl_dict = create_log_lam_grid(dv_temp, wl_min, wl_max)
-        self.wl_final = wl_dict["wl"]
-        self.dv_final = calculate_dv_dict(wl_dict)
-
-        # Create the wl dataset separately using float64 due to rounding errors w/ interpolation.
-        wl_dset = self.hdf5.create_dataset("wl", (len(self.wl_final),), dtype="f8", compression='gzip',
-                                           compression_opts=9)
-        wl_dset[:] = self.wl_final
-        wl_dset.attrs["air"] = self.GridInterface.air
-        wl_dset.attrs["dv"] = self.dv_final
-
-    def process_flux(self, parameters):
-        '''
-        Take a flux file from the raw grid, process it according to the
-        instrument, and insert it into the HDF5 file.
-
-        :param parameters: the model parameters.
-        :type parameters: 1D np.array
-
-        :raises AssertionError: if the `parameters` vector is not
-            the same length as that of the raw grid.
-
-        :returns: a tuple of (parameters, flux, header). If the flux could
-            not be loaded, returns (None, None, None).
-
-        '''
-        # assert len(parameters) == len(config.grid["parname"]), "Must pass numpy array {}".format(config.grid["parname"])
-
-        # If the parameter length is one more than the grid pars,
-        # assume this is for vsini convolution
-        if len(parameters) == (len(config.grid["parname"]) + 1):
-            vsini = parameters[-1]
-            parameters = parameters[:-1]
-        else:
-            vsini = 0.0
-
-        try:
-            flux, header = self.GridInterface.load_flux(parameters)
-
-            # Interpolate the native spectrum to a log-lam FFT grid
-            interp = InterpolatedUnivariateSpline(self.wl_native, flux, k=5)
-            fl = interp(self.wl_FFT)
-            del interp
-            gc.collect()
-
-            # Do the FFT
-            FF = np.fft.rfft(fl)
-
-            if vsini > 0.0:
-                # Calculate the stellar broadening kernel
-                ub = 2. * np.pi * vsini * self.ss
-                sb = j1(ub) / ub - 3 * np.cos(ub) / (2 * ub ** 2) + 3. * np.sin(ub) / (2 * ub ** 3)
-                # set zeroth frequency to 1 separately (DC term)
-                sb[0] = 1.
-
-                # institute vsini and instrumental taper
-                FF_tap = FF * sb * self.taper
-
-            else:
-                # apply just instrumental taper
-                FF_tap = FF * self.taper
-
-            # do IFFT
-            fl_tapered = np.fft.irfft(FF_tap)
-
-            # downsample to the final grid
-            interp = InterpolatedUnivariateSpline(self.wl_FFT, fl_tapered, k=5)
-            fl_final = interp(self.wl_final)
-            del interp
-            gc.collect()
-
-            return (fl_final, header)
-
-        except C.GridError as e:
-            print("No file with parameters {}. C.GridError: {}".format(parameters, e))
-            return (None, None)
-
-    def process_grid(self):
-        '''
-        Run :meth:`process_flux` for all of the spectra within the `ranges`
-        and store the processed spectra in the HDF5 file.
-
-        Only executed in serial for now.
-        '''
-
-        # points is now a list of numpy arrays of the values in the grid
-        # Take all parameter permutations in self.points and create a list
-        # param_list will be a list of numpy arrays, specifying the parameters
-        param_list = []
-
-        # use itertools.product to create permutations of all possible values
-        for i in itertools.product(*self.points):
-            param_list.append(np.array(i))
-
-        all_params = np.array(param_list)
-
-        invalid_params = []
-
-        print("Total of {} files to process.".format(len(param_list)))
-
-        pbar = tqdm(all_params)
-        for i, param in enumerate(pbar):
-            pbar.set_description("Processing {}".format(param))
-            fl, header = self.process_flux(param)
-            if fl is None:
-                print("Deleting {} from all params, does not exist.".format(param))
-                invalid_params.append(i)
-                continue
-
-            # The PHOENIX spectra are stored as float32, and so we do the same here.
-            flux = self.hdf5["flux"].create_dataset(self.key_name.format(*param),
-                                                    shape=(len(fl),), dtype="f", compression='gzip',
-                                                    compression_opts=9)
-            flux[:] = fl
-
-            # Store header keywords as attributes in HDF5 file
-            for key, value in header.items():
-                if key != "" and value != "":  # check for empty FITS kws
-                    flux.attrs[key] = value
-
-        # Remove parameters that do no exist
-        all_params = np.delete(all_params, invalid_params, axis=0)
-
-        par_dset = self.hdf5.create_dataset("pars", all_params.shape, dtype="f8", compression='gzip',
-                                            compression_opts=9)
-        par_dset[:] = all_params
-
-        self.hdf5.close()
-
-
-class HDF5Interface:
-    '''
-    Connect to an HDF5 file that stores spectra.
-    '''
-
-    def __init__(self, filename=config.grid["hdf5_path"], key_name=config.grid["key_name"]):
-        '''
-        :param filename: the name of the HDF5 file
-        :type param: string
-        :param ranges: optionally select a smaller part of the grid to use.
-        :type ranges: dict
-        '''
-        self.filename = os.path.expandvars(filename)
-        self.key_name = key_name
-
-        # In order to properly interface with the HDF5 file, we need to learn
-        # a few things about it
-
-        # 1.) Which parameter combinations exist in the file (self.grid_points)
-        # 2.) What are the minimum and maximum values for each parameter (self.bounds)
-        # 3.) Which values exist for each parameter (self.points)
-
-        with h5py.File(self.filename, "r") as hdf5:
-            self.wl = hdf5["wl"][:]
-            self.wl_header = dict(hdf5["wl"].attrs.items())
-            self.dv = self.wl_header["dv"]
-            self.grid_points = hdf5["pars"][:]
-
-        # determine the bounding regions of the grid by sorting the grid_points
-        low = np.min(self.grid_points, axis=0)
-        high = np.max(self.grid_points, axis=0)
-        self.bounds = np.vstack((low, high)).T
-        self.points = [np.unique(self.grid_points[:, i]) for i in range(self.grid_points.shape[1])]
-
-        self.ind = None  # Overwritten by other methods using this as part of a ModelInterpolator
-
-    def load_flux(self, parameters):
-        '''
-        Load just the flux from the grid, with possibly an index truncation.
-
-        :param parameters: the stellar parameters
-        :type parameters: np.array
-
-        :raises KeyError: if spectrum is not found in the HDF5 file.
-
-        :returns: flux array
-        '''
-
-        key = self.key_name.format(*parameters)
-        with h5py.File(self.filename, "r") as hdf5:
-            try:
-                if self.ind is not None:
-                    fl = hdf5['flux'][key][self.ind[0]:self.ind[1]]
-                else:
-                    fl = hdf5['flux'][key][:]
-            except KeyError as e:
-                raise C.GridError(e)
-
-        # Note: will raise a KeyError if the file is not found.
-
-        return fl
-
-    @property
-    def fluxes(self):
-        '''
-        Iterator to loop over all of the spectra stored in the grid, for PCA.
-
-        Loops over parameters in the order specified by grid_points.
-        '''
-
-        for grid_point in self.grid_points:
-            yield self.load_flux(grid_point)
-
-    def load_flux_hdr(self, parameters):
-        '''
-        Just like load_flux, but also returns the header
-        '''
-        key = self.key_name.format(*parameters)
-        with h5py.File(self.filename, "r") as hdf5:
-            try:
-                hdr = dict(hdf5['flux'][key].attrs)
-                if self.ind is not None:
-                    fl = hdf5['flux'][key][self.ind[0]:self.ind[1]]
-                else:
-                    fl = hdf5['flux'][key][:]
-            except KeyError as e:
-                raise C.GridError(e)
-
-        # Note: will raise a KeyError if the file is not found.
-
-        return (fl, hdr)
 
 
 class IndexInterpolator:
@@ -1133,113 +290,6 @@ class Interpolator:
         fl /= np.median(fl)
         return fl
 
-
-# Convert R to FWHM in km/s by \Delta v = c/R
-class Instrument:
-    '''
-    Object describing an instrument. This will be used by other methods for
-    processing raw synthetic spectra.
-
-    :param name: name of the instrument
-    :type name: string
-    :param FWHM: the FWHM of the instrumental profile in km/s
-    :type FWHM: float
-    :param wl_range: wavelength range of instrument
-    :type wl_range: 2-tuple (low, high)
-    :param oversampling: how many samples fit across the :attr:`FWHM`
-    :type oversampling: float
-    '''
-
-    def __init__(self, name, FWHM, wl_range, oversampling=4.):
-        self.name = name
-        self.FWHM = FWHM  # km/s
-        self.oversampling = oversampling
-        self.wl_range = wl_range
-
-    def __str__(self):
-        '''
-        Prints the relevant properties of the instrument.
-        '''
-        return "Instrument Name: {}, FWHM: {:.1f}, oversampling: {}, " \
-               "wl_range: {}".format(self.name, self.FWHM, self.oversampling, self.wl_range)
-
-
-class TRES(Instrument):
-    '''TRES instrument'''
-
-    def __init__(self, name="TRES", FWHM=6.8, wl_range=(3500, 9500)):
-        super().__init__(name=name, FWHM=FWHM, wl_range=wl_range)
-        # sets the FWHM and wl_range
-
-
-class Reticon(Instrument):
-    '''Reticon Instrument'''
-
-    def __init__(self, name="Reticon", FWHM=8.5, wl_range=(5145, 5250)):
-        super().__init__(name=name, FWHM=FWHM, wl_range=wl_range)
-
-
-class KPNO(Instrument):
-    '''KNPO Instrument'''
-
-    def __init__(self, name="KPNO", FWHM=14.4, wl_range=(6250, 6650)):
-        super().__init__(name=name, FWHM=FWHM, wl_range=wl_range)
-
-
-class SPEX(Instrument):
-    '''SPEX Instrument at IRTF in Hawaii'''
-
-    def __init__(self, name="SPEX", FWHM=150., wl_range=(7500, 54000)):
-        super().__init__(name=name, FWHM=FWHM, wl_range=wl_range)
-
-
-class SPEX_SXD(SPEX):
-    '''SPEX Instrument at IRTF in Hawaii short mode (reduced wavelength range)'''
-
-    def __init__(self, name="SPEX_SXD"):
-        super().__init__(name=name, wl_range=(7500, 26000))
-
-
-class IGRINS(Instrument):
-    '''IGRINS Instruments Abstract Class'''
-
-    def __init__(self, wl_range, name="IGRINS"):
-        super().__init__(name=name, FWHM=7.5, wl_range=wl_range)
-        self.air = False
-
-
-class IGRINS_H(IGRINS):
-    '''IGRINS H band instrument'''
-
-    def __init__(self, name="IGRINS_H", wl_range=(14250, 18400)):
-        super().__init__(name=name, wl_range=wl_range)
-
-class IGRINS_K(IGRINS):
-    '''IGRINS H band instrument'''
-
-    def __init__(self, name="IGRINS_K", wl_range=(18500, 25200)):
-        super().__init__(name=name, wl_range=wl_range)
-
-
-class ESPaDOnS(Instrument):
-    '''ESPaDOnS Instrument'''
-
-    def __init__(self, name="ESPaDOnS", FWHM=4.4, wl_range=(3700, 10500)):
-        super().__init__(name=name, FWHM=FWHM, wl_range=wl_range)
-
-
-class DCT_DeVeny(Instrument):
-    '''DCT DeVeny spectrograph Instrument.'''
-
-    def __init__(self, name="DCT_DeVeny", FWHM=105.2, wl_range=(6000, 10000)):
-        super().__init__(name=name, FWHM=FWHM, wl_range=wl_range)
-
-
-class WIYN_Hydra(Instrument):
-    '''WIYN Hydra spectrograph Instrument.'''
-
-    def __init__(self, name="WIYN_Hydra", FWHM=300., wl_range=(5500, 10500)):
-        super().__init__(name=name, FWHM=FWHM, wl_range=wl_range)
 
 
 def vacuum_to_air(wl):
